+++ src/main/java/de/alpharogroup/xml/json/JsonToObjectExtensions.java
@@ -108,7 +108,7 @@ public final class JsonToObjectExtensions
 	public static <T> T toObject(final @NonNull String jsonString, final @NonNull Class<T> clazz,
 		final boolean newMapper) throws IOException
 	{
-		final ObjectMapper mapper = ObjectMapperFactory.getObjectMapper(newMapper);
+		final ObjectMapper mapper = ObjectMapperFactory.newObjectMapper(newMapper);
 		return toObject(jsonString, clazz, mapper);
 	}
 
@@ -130,7 +130,7 @@ public final class JsonToObjectExtensions
 	public static <T> T toObject(final @NonNull String jsonString, final @NonNull Class<T> clazz,
 		final Module... modules) throws IOException
 	{
-		ObjectMapper mapper = ObjectMapperFactory.getObjectMapper(true);
+		ObjectMapper mapper = ObjectMapperFactory.newObjectMapper(true);
 		mapper = mapper.registerModules(modules);
 		return toObject(jsonString, clazz, mapper);
 	}
@@ -267,7 +267,7 @@ public final class JsonToObjectExtensions
 		@SuppressWarnings("rawtypes") final @NonNull Class<? extends Collection> collectionClass,
 		final @NonNull Class<T> elementClass) throws IOException
 	{
-		final ObjectMapper mapper = ObjectMapperFactory.getObjectMapper(true);
+		final ObjectMapper mapper = ObjectMapperFactory.newObjectMapper(true);
 		return mapper.readValue(jsonString,
 			mapper.getTypeFactory().constructCollectionType(collectionClass, elementClass));
 	}